Added more wrappers for Task methods in RebelTask
